* Modificado el Enum register_cases.py, ahora contienen una cadena descriptiva del error. * Modificado el register_view.py para aprovechar la descripción del enum register_cases.py. * Actualizado el session_controller.py, ahora el método _save_new_user se llama _add_new_user * Creado el metodo read_users() que itera y va retornando los usuarios que se vaya leyendo en el archivo, esto facilita y reduce la repetición de código

+++ src/controllers/user/session_controller.py
@@ -4,6 +4,7 @@
 #
 
 import json
+import os.path
 
 from src.controllers.user.user_controller import password_is_secure, _create_user_from_dict, is_valid_mail
 from src.utils.encrypter import compare_hashed
@@ -12,6 +13,14 @@ from src.enums.register_cases import RegisterCases
 from src.enums.access_level import AccessLevel
 from src.models.users.user import User
 from src.utils.config import REGISTERED_USERS_PATH
+
+def read_users():
+    try:
+        with open(REGISTERED_USERS_PATH, 'r') as file:
+            for user in json.load(file):
+                yield user
+    except (json.JSONDecodeError, FileNotFoundError):
+        return
 
 #metodo en el que se añadirá la lógica de registro de usuarios,
 def register_user(*, fullname:str, username:str, password:str, email: str, access_level:AccessLevel) -> RegisterCases:
@@ -41,7 +50,7 @@ def register_user(*, fullname:str, username:str, password:str, email: str, acces
         "access_level": access_level.name
     }
     new_user = _create_user_from_dict(user_dict)
-    _save_new_user(new_user) # Guardar el usuario en el archivo
+    _add_new_user(new_user) # Guardar el usuario en el archivo
     return RegisterCases.CORRECT
 
 #Metodo en el que se añadirá la logica de inicio de sesión
@@ -49,42 +58,39 @@ def login_user(*, username:str, password:str):
     """Metodo encargado del proceso de login,
     retorna un objeto usuario en caso de un proceso exitoso,
     por el contrario, retornará None"""
-    try:
-        with open(REGISTERED_USERS_PATH, 'r') as file:
-            users = json.load(file)
-            for usr in users: #Iterar entre todos los usuarios registrados
-                if usr.get("username", "") == username:
-                    if compare_hashed(usr.get("password", ""), password): #Compara las contraseñas encriptadas
-                        return _create_user_from_dict(usr)
-            return None
-    except FileNotFoundError:
-        return None
+    for user in read_users():
+        if user["username"] != username:
+            continue
+        else:
+            if compare_hashed(user["password"], password):
+                return _create_user_from_dict(user)
+            else:
+                break
+    return None
 
 def _username_is_taken(username:str):
     """Recibe un username e indica si existe algún usuario registrado con este mismo nombre de usuario"""
-    try:
-        with open(REGISTERED_USERS_PATH, 'r') as file: #Cargar en memoria los usuarios
-            usuarios = json.load(file)
-            for user in usuarios: #Buscar el usuario por su nombre de usuario
-                if user['username'] == username: #Si el nombre de usuario coincide, devolver el usuario
-                    return True
-    except FileNotFoundError: #Si el archivo no existe, devolver None
-        return False
+    for user in read_users():
+        if user["username"] == username:
+            return True
     return False
 
-def _save_new_user(new_user:User):
+def _add_new_user(new_user:User):
     """Recibe un usuario y lo guarda en el archivo local que los almacena"""
-    saved_users = []
     try:
-        with open(REGISTERED_USERS_PATH, 'r') as file: #Cargar en memoria los usuarios
+        if not os.path.exists(REGISTERED_USERS_PATH):
+            with open(REGISTERED_USERS_PATH, 'w') as file:
+                json.dump([], file, indent= 4)
+
+        with open(REGISTERED_USERS_PATH, 'r+') as file: #Cargar en memoria los usuarios
             saved_users = json.load(file)
-    except FileNotFoundError:
-        saved_users = []
-
-    saved_users.append(new_user.to_dict()) #Añadir el nuevo usuario a la lista
-
-    with open(REGISTERED_USERS_PATH, 'w') as file:
-        json.dump(saved_users, file, indent=4) #Guardar la lista de usuarios actualizada
+            saved_users.append(new_user.to_dict())
+            file.seek(0)
+            json.dump(saved_users, file, indent= 4)
+            file.truncate()
+            return True
+    except (FileNotFoundError, json.JSONDecodeError) as e:
+        return False
 
 def _update_user(user_id:str, new_value:dict):
     try:
@@ -94,7 +100,7 @@ def _update_user(user_id:str, new_value:dict):
             for usr in saved_users:
                 if usr.get("id", "") == user_id:
                     for key, value in new_value.items():
-                        if key in usr and key != "id" and key != "password":
+                        if key in usr and key not in {"id", "password"}:
                             usr[key] = value
                             user_found = True
                             break
@@ -102,13 +108,20 @@ def _update_user(user_id:str, new_value:dict):
                 with open(REGISTERED_USERS_PATH, "w") as file:
                     json.dump(saved_users, file, indent=4)
                     return True
-            else:
-                return False
+            return False
     except FileNotFoundError:
         return False
 
 if __name__ == '__main__': #Prueba cerrada
-    user = login_user(username='NuevoUsername',password='Casco12345$')
+    print(register_user(
+        username= 'test_user',
+        password= 'Clavesegura123',
+        access_level= AccessLevel.TECHNICIAN,
+        fullname= 'El Usuario Que Lo Prueba',
+        email= 'correoPaProbar@gmail.com',
+    ))
+
+    user = login_user(username='UsuarioPrueba1',password='Clavesegura123')
 
     if user is not None:
         print(_update_user(user_id= user.id, new_value={"username": 'test_user'}))
@@ -126,4 +139,3 @@ if __name__ == '__main__': #Prueba cerrada
         print('Usuario existe')
         print(type(user))
         print(user.to_dict())
-
